:zap: extract/Change LoadResultHandler to CustomAudioLoadResultHandler class
